Temporarily roll ANGLE forward.
We need some commits which will be available in a future M44 release
that fix the build on Linux when X11 is not present (pure Ozone-Wayland
builds, for example).

More details here:
https://code.google.com/p/angleproject/issues/detail?id=1011

This also allows us to get rid of a temporary patch I had added to fix
the issue only on Tizen.